requirements.txt
pyarrow==14.0.1
adbc_driver_snowflake==0.8.0
python-dotenv==1.0.0